# services/core/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: FeastCloud Core API
  version: 0.1.0
  description: Phase 0 tenant-safe organization, kitchen, order, audit, and edge-inbox API.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organization]
      operationId: listOrganizations
      parameters:
        - $ref: "#/components/parameters/TenantFilter"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/Collection"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Organization]
      operationId: createOrganization
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MutationEnvelope"
                - type: object
                  properties:
                    payload:
                      $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/platform/tenants:
    post:
      tags: [Platform]
      operationId: provisionTenant
      description: >-
        FeastCloud-operator-only onboarding command. It atomically creates a
        customer tenant, organization root, first outlet, first brand, active
        outlet rollout, and template stations. Requires the platform_admin role.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MutationEnvelope"
                - type: object
                  properties:
                    payload: { $ref: "#/components/schemas/TenantProvisionInput" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        default: { $ref: "#/components/responses/Problem" }
...
  /api/v1/orders:
    get:
      tags: [Commerce]
      operationId: listOrders
      parameters:
        - $ref: "#/components/parameters/TenantFilter"
        - $ref: "#/components/parameters/OutletFilter"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200": { $ref: "#/components/responses/Collection" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [Commerce]
      operationId: createOrder
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody: { $ref: "#/components/requestBodies/Mutation" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        default: { $ref: "#/components/responses/Problem" }
...
security:
  - OIDC: []
components:
  securitySchemes:
    OIDC:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
      description: Production mode validates RS256 issuer, audience, tenant, role, outlet, and lifetime claims.
    OutletDeviceCertificate:
      type: mutualTLS
      description: Enrolled outlet device certificate; revocation is checked in PostgreSQL on each sync request.
    DemoTenant:
      type: apiKey
      in: header
      name: X-FeastCloud-Tenant-ID
      description: Phase 0 adapter only; production uses OIDC and device certificates.
    DemoActor:
      type: apiKey
      in: header
      name: X-FeastCloud-Actor-ID
      description: Phase 0 adapter only; production derives the actor from authentication.
  parameters:
...
    TenantFilter:
      name: tenantId
      in: query
      description: Optional assertion; when present it must match the authenticated tenant.
      schema: { type: string, maxLength: 128 }
    OutletFilter:
      name: outletId
      in: query
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 256 }
    CorrelationID:
      name: X-Correlation-ID
      in: header
      schema: { type: string, format: uuid }
  requestBodies:
    Mutation:
      required: true
      content:
        application/json:
          schema:
              $ref: "#/components/schemas/MutationEnvelope"
...
  responses:
    Created:
      description: Resource created. Replays preserve the business result and set Idempotency-Replayed.
      headers:
        Location:
          schema: { type: string }
        Idempotency-Replayed:
          schema: { type: boolean }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ResponseEnvelope" }
...
    Collection:
      description: A cursor-paginated tenant-scoped collection.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CollectionEnvelope" }
    Problem:
      description: Stable machine-readable problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    MutationEnvelope:
      $ref: "../../../packages/contracts/schemas/mutation-envelope.json"
    ResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: { type: string }
            idempotencyReplay: { type: boolean }
    CollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta:
          type: object
          required: [requestId, page]
          properties:
            requestId: { type: string }
            page:
              type: object
              required: [limit]
              properties:
                limit: { type: integer }
                nextCursor: { type: string }
    Problem:
      type: object
      required: [type, title, status, detail, code, messageKey, correlationId, retryable]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        messageKey: { type: string }
        correlationId: { type: string }
        retryable: { type: boolean }
        violations:
          type: array
          items:
            type: object
            required: [field, code, messageKey]
            properties:
              field: { type: string }
              code: { type: string }
              messageKey: { type: string }
...
    OrganizationInput:
      type: object
      additionalProperties: false
      required: [id, name, defaultLocale, defaultCurrency]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 160 }
        legalName: { type: string, maxLength: 240 }
        defaultLocale: { type: string, minLength: 1, maxLength: 35 }
        defaultCurrency: { type: string, pattern: "^[A-Z]{3}$" }
        active: { type: boolean, default: true }
...
    TenantProvisionInput:
      type: object
      additionalProperties: false
      required: [organizationName, ownerName, ownerEmail, defaultLocale, defaultCurrency, outletName, outletCode, timeZone, brandName, brandCode, template]
      properties:
        organizationName: { type: string, minLength: 1, maxLength: 160 }
        legalName: { type: string, maxLength: 240 }
        ownerName: { type: string, minLength: 1, maxLength: 160 }
        ownerEmail: { type: string, format: email, maxLength: 254 }
        defaultLocale: { type: string, maxLength: 35 }
        defaultCurrency: { type: string, pattern: "^[A-Z]{3}$" }
        outletName: { type: string, minLength: 1, maxLength: 160 }
        outletCode: { type: string, minLength: 1, maxLength: 64 }
        timeZone: { type: string }
        brandName: { type: string, minLength: 1, maxLength: 160 }
        brandCode: { type: string, minLength: 1, maxLength: 64 }
        template: { type: string, enum: [restaurant, cloud, central] }
...